+++ DomainManagement/changelog.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## ???
+
+- Upd: Removed most dependencies due to bug in PS5.1. Dependencies in ADMF itself are now expected to provide the necessary tools / modules.
+- Upd: Incremented PSFramework minimum version.
+
 ## 1.4.84 (2020-09-10)
 
 - New: Component: DomainLevel - manage the functional level of your domain