+++ DomainManagement/changelog.md
@@ -4,7 +4,11 @@
 
 - Upd: Removed most dependencies due to bug in PS5.1. Dependencies in ADMF itself are now expected to provide the necessary tools / modules.
 - Upd: Incremented PSFramework minimum version.
+- Upd: Register-DMGroup - added parameter `-Optional` to make a group optional
+- Upd: Renamed test result type `ConfigurationOnly` to `Create`
+- Upd: Test-DMGroup - don't create Create actions for groups that are optional
 - Fix: Test-GPPermission - path-based filters would not correctly map permissions
+- Fix: Component Object - Does not allow modifying properties on the domain object itself
 
 ## 1.4.84 (2020-09-10)